requirements.txt
ambient-api==1.5.6
ambient-aprs==0.13.0
asgiref==3.5.2
ask-sdk-core==1.17.1
ask-sdk-model==1.34.2
ask-sdk-runtime==1.17.1
ask-sdk-webservice-support==1.3.3
asn1crypto==1.5.1
Babel==2.11.0
beautifulsoup4==4.11.1
bleach==5.0.1
CacheControl==0.12.11
certifi==2022.12.7
certvalidator==0.11.1
contractions
cffi==1.15.1
chardet==5.0.0
charset-normalizer
cryptography
deprecation==2.1.0
Django==4.1.6
django-admin-sortable==2.2.4
django-allauth
django-allauth-templates-bootstrap4
django-ask-sdk==1.0.0
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-extensions==3.2.1
django-filer==2.2.3
django-geojson==4.0.0
django-js-asset==2.0.0
django-leaflet==0.28.3
django-mptt==0.14.0
django-polymorphic==3.1.0
djangorestframework==3.14.0
duckduckgo_search
easy-thumbnails==2.8.3
ecdsa==0.18.0
et-xmlfile==1.1.0
favorites-icons==1.0.7
freezegun==1.2.2
frontend-assets==0.8.1
future==0.18.2
geo-ez==0.1.5
geolite2utils
holidays
idna==3.3
iso18245==1.2.0
jsonpickle
lxml==4.9.2
msgpack==1.0.4
nltk
openai
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
phonetisch==0.2.2
pika==1.3.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pybase62==0.5.0
pycparser==2.21
pynws
pyotp==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
python-jose==3.3.0
python-Levenshtein==0.20.8
python-u2flib-server==5.0.1
pytz==2022.6
pyxlsb==1.0.10
reportlab==3.6.12
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
squareup
ssml-builder==1.1.1
starkbank-ecdsa==2.2.0
subresource-integrity==0.2
svglib==1.4.1
tinycss2==1.2.1
tzdata==2022.7
ua-parser==0.16.1
Unidecode
urllib3==1.26.13
user-agents==2.2.0
uuid0==0.2.7
webencodings==0.5.1
xmltodict==0.13.0
